+++ tapo-py/src/requests.rs
@@ -1,7 +1,9 @@
 mod energy_data_interval;
 mod play_alarm;
+mod power_data_interval;
 mod set_device_info;
 
 pub use energy_data_interval::*;
 pub use play_alarm::*;
+pub use power_data_interval::*;
 pub use set_device_info::*;