+++ tapo-py/src/requests/power_data_interval.rs
@@ -1,0 +1,8 @@
+use pyo3::prelude::*;
+
+#[derive(Clone, PartialEq)]
+#[pyclass(name = "PowerDataInterval", eq, eq_int)]
+pub enum PyPowerDataInterval {
+    Every5Minutes,
+    Hourly,
+}