+++ tapo/examples/tapo_p110.rs
@@ -1,10 +1,10 @@
 /// P110, P110M and P115 Example
 use std::{env, thread, time::Duration};
 
-use chrono::{DateTime, Datelike, NaiveDate, Utc};
+use chrono::{Datelike, NaiveDate, Utc};
 use log::info;
 use tapo::ApiClient;
-use tapo::requests::EnergyDataInterval;
+use tapo::requests::{EnergyDataInterval, PowerDataInterval};
 
 mod common;
 
@@ -41,26 +41,20 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let energy_usage = device.get_energy_usage().await?;
     info!("Energy usage: {energy_usage:?}");
 
-    let current_date = Utc::now();
+    let current_date = Utc::now().naive_utc().date();
 
+    // Energy data - Hourly interval
+    // `start_date` and `end_date` are an inclusive interval that must not be greater than 8 days.
     let energy_data_hourly = device
         .get_energy_data(EnergyDataInterval::Hourly {
-            start_date: NaiveDate::from_ymd_opt(
-                current_date.year(),
-                current_date.month(),
-                current_date.day(),
-            )
-            .unwrap(),
-            end_date: NaiveDate::from_ymd_opt(
-                current_date.year(),
-                current_date.month(),
-                current_date.day(),
-            )
-            .unwrap(),
+            start_date: current_date,
+            end_date: current_date,
         })
         .await?;
     info!("Energy data (hourly): {energy_data_hourly:?}");
 
+    // Energy data - Daily interval
+    // `start_date` must be the first day of a quarter.
     let energy_data_daily = device
         .get_energy_data(EnergyDataInterval::Daily {
             start_date: NaiveDate::from_ymd_opt(
@@ -73,6 +67,8 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         .await?;
     info!("Energy data (daily): {energy_data_daily:?}");
 
+    // Energy data - Monthly interval
+    // `start_date` must be the first day of a year.
     let energy_data_monthly = device
         .get_energy_data(EnergyDataInterval::Monthly {
             start_date: NaiveDate::from_ymd_opt(current_date.year(), 1, 1).unwrap(),
@@ -80,14 +76,45 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         .await?;
     info!("Energy data (monthly): {energy_data_monthly:?}");
 
+    // Power data - Every 5 minutes interval
+    // `start_date_time` and `end_date_time` describe an exclusive interval.
+    // If the result would yield more than 144 entries (i.e. 12 hours),
+    // the `end_date_time` will be adjusted to an earlier date and time.
+    let power_data_every_5_minutes = device
+        .get_power_data(PowerDataInterval::Every5Minutes {
+            start_date_time: Utc::now() - chrono::Duration::hours(12),
+            end_date_time: Utc::now(),
+        })
+        .await?;
+    info!(
+        "Power data (every 5 minutes): Start date time '{}', End date time '{}', Entries {}, First entry: {:?}",
+        power_data_every_5_minutes.start_date_time,
+        power_data_every_5_minutes.end_date_time,
+        power_data_every_5_minutes.entries.len(),
+        power_data_every_5_minutes.entries.first()
+    );
+
+    // Power data - Hourly interval
+    // `start_date_time` and `end_date_time` describe an exclusive interval.
+    // If the result would yield more than 144 entries (i.e. 6 days),
+    // the `end_date_time` will be adjusted to an earlier date and time.
+    let power_data_hourly = device
+        .get_power_data(PowerDataInterval::Hourly {
+            start_date_time: Utc::now() - chrono::Duration::days(3),
+            end_date_time: Utc::now(),
+        })
+        .await?;
+    info!(
+        "Power data (hourly): Start date time '{}', End date time '{}', Entries {}, First entry: {:?}",
+        power_data_hourly.start_date_time,
+        power_data_hourly.end_date_time,
+        power_data_hourly.entries.len(),
+        power_data_hourly.entries.first()
+    );
+
     Ok(())
 }
 
-fn get_quarter_start_month(current_date: &DateTime<Utc>) -> u32 {
-    match current_date.month() {
-        m if m < 4 => 1,
-        m if m < 7 => 4,
-        m if m < 10 => 7,
-        _ => 10,
-    }
+fn get_quarter_start_month(current_date: &NaiveDate) -> u32 {
+    ((current_date.month() - 1) / 3) * 3 + 1
 }