+++ tapo/src/responses.rs
@@ -11,6 +11,7 @@ mod device_usage_result;
 mod energy_data_result;
 mod energy_usage_result;
 mod handshake_result;
+mod power_data_result;
 mod supported_alarm_type_list_result;
 mod tapo_response;
 mod tapo_result;
@@ -27,6 +28,7 @@ pub use device_usage_energy_monitoring_result::*;
 pub use device_usage_result::*;
 pub use energy_data_result::*;
 pub use energy_usage_result::*;
+pub use power_data_result::*;
 pub use trigger_logs_result::*;
 
 pub(crate) use control_child_result::*;